create a dedicated HttpUpgradesTest
